+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-# 0.8.0
+# 1.0.0
 - Fully supports Enka.Network API.
 - Added User#getCharacters().
 - Added User#enkaUserHash.
+- Added ClientOptions#enkaSystem.
+- Renamed User to StarRailUser.
 # 0.7.6
 - Added ClientOptions#apiBaseUrl for custom api base url such as `https://enka.network/api/hsr/uid`.
 # 0.7.5